Group the models by requests
