+++ sdk-core/src/main/java/com/reach5/identity/sdk/core/JavaReachFive.kt
@@ -3,6 +3,7 @@ package com.reach5.identity.sdk.core
 import android.app.Activity
 import android.content.Intent
 import com.reach5.identity.sdk.core.models.*
+import com.reach5.identity.sdk.core.models.requests.UpdatePasswordRequest
 import com.reach5.identity.sdk.core.utils.Callback
 
 class JavaReachFive(activity: Activity, sdkConfig: SdkConfig, providersCreators: List<ProviderCreator>) {
