+++ sdk-core/src/main/java/com/reach5/identity/sdk/core/models/requests/RequestPasswordResetRequest.kt
@@ -1,0 +1,16 @@
+package com.reach5.identity.sdk.core.models.requests
+
+import android.os.Parcelable
+import com.google.gson.annotations.SerializedName
+import kotlinx.android.parcel.Parcelize
+
+@Parcelize
+data class RequestPasswordResetRequest(
+    @SerializedName("client_id")
+    val clientId: String,
+    val email: String?,
+    @SerializedName("redirect_url")
+    val redirectUrl: String? = null,
+    @SerializedName("phone_number")
+    val phoneNumber: String?
+) : Parcelable
