+++ sdk-core/src/main/java/com/reach5/identity/sdk/core/models/requests/UpdateEmailRequest.kt
@@ -1,0 +1,12 @@
+package com.reach5.identity.sdk.core.models.requests
+
+import android.os.Parcelable
+import com.google.gson.annotations.SerializedName
+import kotlinx.android.parcel.Parcelize
+
+@Parcelize
+data class UpdateEmailRequest(
+    val email: String,
+    @SerializedName("redirect_url")
+    val redirectUrl: String? = null
+) : Parcelable
