+++ sdk-core/src/main/java/com/reach5/identity/sdk/core/models/requests/UpdatePhoneNumberRequest.kt
@@ -1,0 +1,11 @@
+package com.reach5.identity.sdk.core.models.requests
+
+import android.os.Parcelable
+import com.google.gson.annotations.SerializedName
+import kotlinx.android.parcel.Parcelize
+
+@Parcelize
+data class UpdatePhoneNumberRequest(
+    @SerializedName("phone_number")
+    val phoneNumber: String
+) : Parcelable
