+++ sdk-core/src/main/java/com/reach5/identity/sdk/core/models/requests/VerifyPhoneNumberRequest.kt
@@ -1,0 +1,13 @@
+package com.reach5.identity.sdk.core.models.requests
+
+import android.os.Parcelable
+import com.google.gson.annotations.SerializedName
+import kotlinx.android.parcel.Parcelize
+
+@Parcelize
+data class VerifyPhoneNumberRequest(
+    @SerializedName("phone_number")
+    val phoneNumber: String,
+    @SerializedName("verification_code")
+    val verificationCode: String
+) : Parcelable
