+++ sdk-facebook/src/main/java/com/reach5/identity/sdk/facebook/FacebookProvider.kt
@@ -13,6 +13,7 @@ import com.reach5.identity.sdk.core.ProviderCreator
 import com.reach5.identity.sdk.core.api.ReachFiveApi
 import com.reach5.identity.sdk.core.api.ReachFiveApiCallback
 import com.reach5.identity.sdk.core.models.*
+import com.reach5.identity.sdk.core.models.requests.LoginProviderRequest
 import com.reach5.identity.sdk.core.utils.Failure
 import com.reach5.identity.sdk.core.utils.Success
